feat: AI 개선 확인 다이얼로그 및 일일 토큰 사용량 영속화
- AIConfirmationOverlay 추가로 AI 개선 실행 전 확인 단계 도입
- AITokenDailyUsageRecord(SwiftData) 모델로 멤버별 당일 누적 토큰 저장/복원
- 본문 공백/AI 처리 중일 때 툴바 AI 버튼 비활성화
- AILoadingOverlay에서 토큰 사용량 표시 제거(확인 다이얼로그로 이동) 및 문구 정리
- NoticeRichTextView를 Coordinator 확장·Representable 파일로 모듈 분리
