+++ AppProduct/AppProduct/Features/Notice/Data/AITokenDailyUsageRecord.swift
@@ -1,0 +1,35 @@
+//
+//  AITokenDailyUsageRecord.swift
+//  AppProduct
+//
+//  Created by euijjang97 on 4/18/26.
+//
+
+import Foundation
+import SwiftData
+
+/// AI 토큰 일일 사용량 로컬 저장 모델 (SwiftData + CloudKit Sync)
+///
+/// 멤버별 당일 AI 개선 기능 토큰 누적 사용량을 보존합니다.
+/// 키는 (memberId, date의 startOfDay) 조합으로 식별합니다.
+@Model
+final class AITokenDailyUsageRecord {
+
+    // MARK: - Property
+
+    var memberId: Int = 0
+    var date: Date = Date()
+    var usedTokens: Int = 0
+
+    // MARK: - Init
+
+    init(
+        memberId: Int,
+        date: Date,
+        usedTokens: Int
+    ) {
+        self.memberId = memberId
+        self.date = date
+        self.usedTokens = usedTokens
+    }
+}
