+++ AppProduct/AppProduct/Features/Notice/Presentation/Views/NoticeEditor/NoticeEditorPresentations.swift
@@ -32,11 +32,25 @@ struct NoticeEditorPresentations: ViewModifier {
             .fullScreenCover(isPresented: $viewModel.showVoting, content: votingSheet)
             .alertPrompt(item: $viewModel.alertPrompt)
             .overlay {
+                if viewModel.showAIConfirmation {
+                    AIConfirmationOverlay(
+                        tokenUsage: viewModel.aiTokenUsage,
+                        onConfirm: {
+                            Task { await viewModel.startAIImprovement() }
+                        },
+                        onCancel: {
+                            viewModel.showAIConfirmation = false
+                        }
+                    )
+                    .transition(.opacity.combined(with: .scale(scale: 0.94, anchor: .center)))
+                }
+            }
+            .animation(.easeInOut(duration: 0.25), value: viewModel.showAIConfirmation)
+            .overlay {
                 if viewModel.isAIProcessing || viewModel.showAICompletionSummary {
                     AILoadingOverlay(
                         phase: viewModel.isAIProcessing ? .processing : .completed,
                         streamingText: viewModel.aiStreamingText,
-                        tokenUsage: viewModel.aiTokenUsage,
                         onConfirm: { viewModel.dismissAICompletionSummary() }
                     )
                     .transition(.opacity.combined(with: .scale(scale: 0.94, anchor: .center)))
