+++ AppProduct/AppProduct/Features/Notice/Presentation/Views/NoticeEditor/RichTextEditor/NoticeRichText/UITextView+MinimumHeight.swift
@@ -1,0 +1,14 @@
+//
+//  UITextView+MinimumHeight.swift
+//  AppProduct
+//
+
+import UIKit
+
+extension UITextView {
+
+    var minimumHeight: CGFloat {
+        let lineHeight = font?.lineHeight ?? UIFont.preferredFont(forTextStyle: .body).lineHeight
+        return ceil(lineHeight + textContainerInset.top + textContainerInset.bottom)
+    }
+}
